refactor: move all files inside "yesman.epicfight.kubejs" to "yesman.epicfight.compat.kubejs" using IDEA refactoring
